Verify image

+++ client/src/scenes/widgets/CommentSection.jsx
@@ -8,15 +8,16 @@ import {
   Avatar,
   IconButton,
 } from "@mui/material";
+import { DeleteOutlined } from "@mui/icons-material";
 import { useDispatch, useSelector } from "react-redux";
 import { setPost } from "state";
-import DeleteComment from "components/DeleteComment"; // Adjust the path as needed
 
 const CommentSection = ({ postId, comments }) => {
   const [newComment, setNewComment] = useState("");
   const dispatch = useDispatch();
   const token = useSelector((state) => state.token);
   const loggedInUserId = useSelector((state) => state.user._id);
+  const [isLoading, setIsLoading] = useState(false); 
 
   const handleAddComment = async () => {
     const response = await fetch(`http://localhost:3001/posts/comment`, {
@@ -36,21 +37,34 @@ const CommentSection = ({ postId, comments }) => {
   };
 
   const handleDeleteComment = async (commentId) => {
-    const response = await fetch(
-      `http://localhost:3001/posts/${postId}/comments/${commentId}`,
-      {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${token}`, // Include your auth token in headers
-        },
-      }
-    );
+    const confirmation = window.confirm("Are you sure you want to delete this comment?");
 
-    if (response.ok) {
-      const updatedPost = await response.json();
-      dispatch(setPost({ post: updatedPost }));
+    if (confirmation) {
+      setIsLoading(true); // Set loading state (optional)
+      try {
+        const response = await fetch(
+          `http://localhost:3001/posts/${postId}/comments/${commentId}`,
+          {
+            method: "DELETE",
+            headers: {
+              Authorization: `Bearer ${token}`, // Include your auth token in headers
+            },
+          }
+        );
+    
+        if (response.ok) {
+          const updatedPost = await response.json();
+          dispatch(setPost({ post: updatedPost }));
+        } else {  
+          console.error("Failed to delete comment");
+        }
+      } catch (error) {
+        console.error("Error deleting post:", error);
+      } finally {
+        setIsLoading(false); // Reset loading state
+      }
     } else {
-      console.error("Failed to delete comment");
+      console.log("Comment deletion cancelled");
     }
   };
 
@@ -74,8 +88,8 @@ const CommentSection = ({ postId, comments }) => {
               </Typography>
             </Box>
             {comment.userId === loggedInUserId && (
-              <IconButton onClick={() => handleDeleteComment(comment._id)}>
-                <DeleteComment postId={postId} commentId={comment._id} />
+              <IconButton disabled={isLoading} sx={{ color: "error.main"}} onClick={() => handleDeleteComment(comment._id)}>
+                <DeleteOutlined />
               </IconButton>
             )}
           </Box>
